refactor: update minimum SDK version to 19 and rename NativeRNBridge to NativeBridge in imports
