+++ libs/sdk/android/gradle.properties
@@ -1,5 +1,5 @@
 RNRudderSdkModule_kotlinVersion=2.0.21
-RNRudderSdkModule_minSdkVersion=24
+RNRudderSdkModule_minSdkVersion=19
 RNRudderSdkModule_targetSdkVersion=34
 RNRudderSdkModule_compileSdkVersion=35
 RNRudderSdkModule_ndkVersion=27.1.12297006
